Add the injected preferences to the parent of the PreferenceGroup
Also,update the order field.

When preferences are injected into a LogicalPreferenceGroup the
androidx.PreferenceAdapter will flatten the hierarchy. But the
LogicalPreferenceGroup is not set to invisible and will be added
to the list of visible items. If no preferences are injected into
a LogicalPreferenceGroup, the group will be automatically set to
invisible.

This change will add the injected preference automatically to the
parent of the LogicalPreferenceGroup as it would be done by the
PreferenceAdapter. In this case the LogicalPreferenceGroup is still
empty and will be automatically set to invisible.

Changing also the homepage_fragment to show usage.

Bug: 225173698
Test: manual and atest CarSettingsUnitTests
Change-Id: I2c35bd60ba16449df66ee123d51f17929adf9b43