Add required flag to receiver in LRAPC
Android T allows apps to declare a runtime receiver as not exported
by invoking registerReceiver with a new RECEIVER_NOT_EXPORTED flag;
receivers registered with this flag will only receive broadcasts from
the platform and the app itself. However to ensure developers can
properly protect their receivers, all apps targeting U
and registering a receiver for non-system broadcasts must
specify either the exported or not exported flag when invoking
registerReceiver; if one of these flags is not provided, the
platform will throw a SecurityException. This commit updates
the receiver with the RECEIVER_EXPORTED flag since the
broadcast is sent from outside the app.

Bug: 234659204
Test: Build
Change-Id: Ibc059d6734b5ab3f8917c0ee4ccd9a8adc9db736