Fix lockscreen timeout not being cleared
The timer is cleared when the activity is paused and relies on checking
the status of the timeout during onResume(). However, the timeout
message was never cleared with the timer, so if there is no more timeout
in onResume() the message will not get updated.

Bug: 240328288
Test: Manual and atest
com.android.car.settings.security.ConfirmLockLockoutHelperTest

Change-Id: I455c67584361f51577676fdc16c4868740d00b7f