Simplify lockscreen credential validation in InitialLockSetupService
PasswordHelper is the only remaining caller of
PasswordMetrics#validatePassword(). validatePassword() can't correctly
check for invalid characters, and it only works for credentials of type
PIN and PASSWORD. Its replacement is validateCredential() which takes a
LockscreenCredential argument.

Therefore, this CL updates PasswordHelper to consistently take a
LockscreenCredential in its validation methods and pass that on to
PasswordMetrics#validateCredential(). For that to work, it updates
InitialLockSetupService to create a LockscreenCredential for the (int
lockType, byte[] password) pair that its methods take.

Note that PasswordHelper previously had three validation methods:
validate(), validateSetupWizard(), and
validateSetupWizardAndReturnError(). These are unified into two.

There shouldn't be any user visible behavior changes as a result of this
cleanup. However, this is a prerequisite for correctly rejecting
non-ASCII characters in lockscreen passwords.

Bug: 219511761
Bug: 232900169
Bug: 243881358
Test: atest CarSettingsUnitTests:{PasswordHelperTest,ChooseLockPinPasswordFragmentTest}
      (I also wanted to run InitialLockSetupServiceTest, but I can't
      find a way to get that test to run. Many of the assertions done in
      that test also can't possibly be passing currently.)
Change-Id: I1f3c7b5a42b747eb371a1811e12756e2b3952498