Fix CarUiTwoActionSwitchPreference test crashes
Chassis made a change that required onBind() to be called before
onClick() in CarUiTwoActionSwitchPreference. Rather than set up
activities for all of our unit tests, we are going with the
approach of mocking the preference since we are only care about
the logic of the PreferenceController.

Bug: 283075014
Test: atest
com.android.car.settings.notifications.NotificationsAppListPreferenceControllerTest
and atest com.android.car.settings.network.MobileNetworkEntryPreferenceControllerTest

Change-Id: I829770d8e96921fc3c834b9204fcb270150d35c3