Exclude ephemeral user from getAllAdminProfiles().
An ephemeral user is set to be removed as soon as it switches to the
background. When an admin user is being removed, there is a check if it
is the last admin user, to make sure there is at least one remaining
admin user. Such check is currently not working correctly when the
previous admin user is set to ephemeral.

This CL fixes the issue by excluding ephemeral users when counting the
number of admin users.

Bug: 304720114

Test: Manually with emulator
Change-Id: Id3f689f5e58cb24d1561abefb6b4da4a8a1649c9