Skip profile preferences that are not meant to be shown on the UI
Problem: Profiles like VCP and CSIP do not provide title resource IDs,
as they're framework profiles for LE Audio and are not meant to be shown
on the UI on their own. Despite this, our Car Settings code recognizes
their existence and fails to filter them out. This causes a crash when
the resource ID of 0x0 is used to load their title.

Solution: Do what the phones do and leverage accessProfileEnabled() to
skip over LocalBluetoothProfiles that are not supposed to be shown on
the UI.

https://screencast.googleplex.com/cast/NjY0NDQ5MTkxNzE5NzMxMnxmMDRkOWEwNS1lZA

Flag: EXEMPT, trivial bug fix
Bug: 422776267
Test: m -j
Change-Id: I499cf7b10645309bb8540f2aa1ea574bc26ac797