Mock accessProfileEnabled in tests
Problem: We're now using the accessProfileEnabled() function to
determine if we should display a given profile preference. This is in
parity with mobile. But, our tests were not updated to mock this. This
meant that the default function return value for an unmocked function
was used, which is false. This meant our mocked profiles were filtered
out, and tests were failing

Solution: Mock accessProfileEnabled() to true. We can further extract
device mocking to a function so its easy to keep all tests consistent
and future changes can be made in one place.

Flag: EXEMPT, test only
Bug: 423669669
Test: atest BluetoothDeviceProfilesPreferenceControllerTest
Change-Id: Ifb9bf63db6073790cbe501e804143cdd885286c6