Fix CarAspectRatioDialogActivityTest failures
This commit fixes the CarAspectRatioDialogActivityTest failures by resolving a Mockito
session conflict and registering the CarAspectRatioDialogActivity in the
test app's AndroidManifest.xml. The Mockito UnfinishedMockingSessionException
was addressed by removing the redundant MockitoRule and relying solely on
ExtendedMockito for static method mocking. Additionally, the `Unable to resolve
activity for: Intent error was fixed by declaring CarAspectRatioDialogActivity`
in the AndroidManifest.xml and using a locally available theme.

Bug: 450990258
Test: atest CarSettingsMultivalentDeviceTests:CarAspectRatioDialogActivityTest
Flag: EXEMPT TEST_ONLY

Change-Id: I02b77213c8f04b92afb29a6c088b2742d4e7ea07