Update tests for AudioRouteItem.State enum
Update tests for the `AudioRouteItem.State` enum to align with recent modifications.

A new test is added to verify the integer value for each state, ensuring the stability of these values which may be used for serialization. Existing tests are also updated to include assertions for the newly added `UNSPECIFIED` state.



This CL is generated by AI, please check the comments for more details.

The purpose of this change is to increase test coverage for the code affected by the source change that you recently reviewed or merged: I64942bda4cfcfdd44d7daab824b3578e32f65600

Test: abtd test passed go/forrest-run/L19000030046004037
Bug: 431235865 (Umbrella)
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true



Change-Id: Iafa8dc8afea5ae703517707cf487f44f3254fea6